move help content to component
