+++ apps/studio/components/layouts/ProjectLayout/LayoutHeader/HelpDropdown/HelpContent.tsx
@@ -1,0 +1,85 @@
+import { IS_PLATFORM } from 'common'
+import type { SupportFormUrlKeys } from 'components/interfaces/Support/SupportForm.utils'
+import { SIDEBAR_KEYS } from 'components/layouts/ProjectLayout/LayoutSidebar/LayoutSidebarProvider'
+import { useRouter } from 'next-router-mock'
+import Image from 'next/image'
+import SVG from 'react-inlinesvg'
+import { useAiAssistantStateSnapshot } from 'state/ai-assistant-state'
+import { useSidebarManagerSnapshot } from 'state/sidebar-manager-state'
+import { Button, cn, Separator } from 'ui'
+import styleHandler from 'ui/src/lib/theme/styleHandler'
+
+import { ASSISTANT_SUGGESTIONS } from './HelpDropdown.constants'
+import { HelpSection } from './HelpSection'
+
+export const HelpContent = ({
+  onClose,
+  projectRef,
+  supportLinkQueryParams,
+}: {
+  onClose: () => void
+  projectRef: string | undefined
+  supportLinkQueryParams: Partial<SupportFormUrlKeys> | undefined
+}) => {
+  const snap = useAiAssistantStateSnapshot()
+  const { openSidebar } = useSidebarManagerSnapshot()
+  const router = useRouter()
+
+  const __styles = styleHandler('popover')
+
+  return (
+    <>
+      <HelpSection
+        className="px-5"
+        excludeIds={['discord']}
+        isPlatform={IS_PLATFORM}
+        projectRef={projectRef}
+        supportLinkQueryParams={supportLinkQueryParams}
+        onAssistantClick={() => {
+          onClose()
+          openSidebar(SIDEBAR_KEYS.AI_ASSISTANT)
+          snap.newChat(ASSISTANT_SUGGESTIONS)
+        }}
+        onSupportClick={onClose}
+      />
+      <Separator className={cn(__styles.separator, 'my-4')} />
+      <div className="flex flex-col gap-4">
+        <div className="px-5 flex flex-col gap-0.5">
+          <h5 className="text-foreground">Community support</h5>
+          <p className="text-xs text-foreground-lighter text-balance">
+            Our Discord community can help with code-related issues. Many questions are answered in
+            minutes.
+          </p>
+        </div>
+        <div className="px-5">
+          <div
+            className="relative space-y-2 overflow-hidden rounded px-5 py-4 pb-12 shadow-md"
+            style={{ background: '#404EED' }}
+          >
+            <a
+              href="https://discord.supabase.com"
+              target="_blank"
+              rel="noreferrer"
+              className="group dark block cursor-pointer"
+            >
+              <Image
+                className="absolute left-0 top-0 opacity-50 transition-opacity group-hover:opacity-40"
+                src={`${router.basePath}/img/support/discord-bg-small.jpg`}
+                layout="fill"
+                objectFit="cover"
+                alt="Discord illustration"
+              />
+              <Button
+                type="secondary"
+                size="tiny"
+                icon={<SVG src={`${router.basePath}/img/discord-icon.svg`} className="h-4 w-4" />}
+              >
+                <span style={{ color: '#404EED' }}>Join us on Discord</span>
+              </Button>
+            </a>
+          </div>
+        </div>
+      </div>
+    </>
+  )
+}
